+++ test/components/uchi/field/has_many_test.rb
@@ -20,7 +20,7 @@ module Uchi
       end
 
       test "has custom collection_query" do
-        custom_query = ->(query) { query.where(published: true) }
+        custom_query = ->(query:) { query.where(published: true) }
         field = Uchi::Field::HasMany.new(:titles).collection_query(custom_query)
         assert_equal custom_query, field.collection_query
       end